+++ Platform.App/python-template/app/blueprints/mapping.py
@@ -1,34 +1,16 @@
-""" Server API """
 import json
-from flask import Flask, request, jsonify
-from api.query_controller import QueryController
-from api.command_controller import CommandController
+
+from flask import Blueprint, request, jsonify
+
 from mapper.builder import MapBuilder
-from app.query_service import QueryService
-from settings.loader import Loader as SettingsLoader
-from database import create_session
-
-from flask import request
+from app.services import QueryService
+from app.controllers import QueryController, CommandController
 
 
-env = SettingsLoader().load()
-app = Flask(__name__, instance_relative_config=True)
-app.debug = False
-app.session_factory = create_session
+mapping = Blueprint('simple_page', __name__)
 
 
-@app.before_request
-def create_request_session():
-    setattr(request, 'session', app.session_factory())
-
-
-@app.after_request
-def destroy_request_session(response):
-    request.session.close()
-    return response
-
-
-@app.route("/<app_id>/<entity>", methods=['GET'])
+@mapping.route("/<app_id>/<entity>", methods=['GET'])
 def query_map(app_id, entity):
     """ Query data on domain """
     try:
@@ -48,7 +30,7 @@ def query_map(app_id, entity):
         return jsonify(resp), resp["code"]
 
 
-@app.route("/<app_id>/persist", methods=['POST'])
+@mapping.route("/<app_id>/persist", methods=['POST'])
 def persist_map(app_id):
     """ Persist data on domain """
     instance_id = request.headers.get('Instance-Id')
