+++ Platform.App/python-template/app/controllers/__init__.py
@@ -1,0 +1,2 @@
+from .command_controller import CommandController
+from .query_controller import QueryController
