+++ Platform.App/python-template/app/controllers/command_controller.py
