+++ Platform.App/python-template/app/controllers/query_controller.py
