+++ Platform.App/python-template/app/services/__init__.py
@@ -1,0 +1,2 @@
+from .query_service import QueryService
+
