+++ Platform.App/python-template/app/services/query_service.py
