+++ Platform.App/python-template/app/tests/controllers/test_command_controller.py
