+++ Platform.App/python-template/app/tests/controllers/test_query_controller.py
