+++ Platform.App/python-template/main.py
@@ -1,5 +1,7 @@
 ''' Main Module starts all domain app components '''
-from migration.sync import sync_db, wait_postgres
-if wait_postgres():
-    sync_db()
+from migration.sync import wait_postgres
+from database import create_db
 
+
+if wait_postgres():
+    create_db()
